+++ R/comparisons.R
@@ -81,6 +81,7 @@
 #' @template comparison_functions
 #' @template bayesian
 #' @template equivalence
+#' @template references
 #'
 #' @return A `data.frame` with one row per observation (per term/group) and several columns:
 #' * `rowid`: row number of the `newdata` data frame
@@ -90,6 +91,9 @@
 #' * `dydx`: slope of the outcome with respect to the term, for a given combination of predictor values
 #' * `std.error`: standard errors computed by via the delta method.
 #' * `p.value`: p value associated to the `estimate` column. The null is determined by the `hypothesis` argument (0 by default), and p values are computed before applying the `transform` argument.
+#' * `s.value`: Shannon information tranforms of p values. How many consecutive "heads" tosses would provide the same amount of evidence (or "suprise") against the null hypothesis that the coin is fair? See Greenland (2019) and Cole et al. (2020).
+#' * `conf.low`: lower bound of the confidence interval (or equal-tailed interval for bayesian models)
+#' * `conf.high`: upper bound of the confidence interval (or equal-tailed interval for bayesian models)
 #'
 #' See `?print.marginaleffects` for printing options.
 #'
@@ -532,7 +536,7 @@ comparisons <- function(model,
         "rowid", "rowidcf", "group", "term", "hypothesis", "by",
         grep("^contrast", colnames(mfx), value = TRUE),
         bycols,
-        "estimate", "std.error", "statistic", "p.value", "conf.low",
+        "estimate", "std.error", "statistic", "p.value", "s.value", "conf.low",
         "conf.high", "df", "predicted", "predicted_hi", "predicted_lo")
     cols <- intersect(stubcols, colnames(mfx))
     cols <- unique(c(cols, colnames(mfx)))