+++ R/slopes.R
@@ -106,6 +106,7 @@
 #' @template model_specific_arguments
 #' @template bayesian
 #' @template equivalence
+#' @template references
 #'
 #' @return A `data.frame` with one row per observation (per term/group) and several columns:
 #' * `rowid`: row number of the `newdata` data frame
@@ -114,6 +115,10 @@
 #' * `term`: the variable whose marginal effect is computed
 #' * `dydx`: slope of the outcome with respect to the term, for a given combination of predictor values
 #' * `std.error`: standard errors computed by via the delta method.
+#' * `p.value`: p value associated to the `estimate` column. The null is determined by the `hypothesis` argument (0 by default), and p values are computed before applying the `transform` argument. For models of class `feglm`, `Gam`, `glm` and `negbin`, p values are computed on the link scale by default unless the `type` argument is specified explicitly.
+#' * `s.value`: Shannon information tranforms of p values. How many consecutive "heads" tosses would provide the same amount of evidence (or "suprise") against the null hypothesis that the coin is fair? See Greenland (2019) and Cole et al. (2020).
+#' * `conf.low`: lower bound of the confidence interval (or equal-tailed interval for bayesian models)
+#' * `conf.high`: upper bound of the confidence interval (or equal-tailed interval for bayesian models)
 #'
 #' See `?print.marginaleffects` for printing options.
 #'
@@ -277,7 +282,7 @@ slopes <- function(model,
     data.table::setDT(out)
 
     # clean columns
-    stubcols <- c("rowid", "group", "term", "contrast", "hypothesis", "dydx", "estimate", "std.error", "statistic", "p.value", "conf.low", "conf.high",
+    stubcols <- c("rowid", "group", "term", "contrast", "hypothesis", "dydx", "estimate", "std.error", "statistic", "p.value", "s.value", "conf.low", "conf.high",
                   sort(grep("^predicted", colnames(newdata), value = TRUE)))
     cols <- intersect(stubcols, colnames(out))
     cols <- unique(c(cols, colnames(out)))